backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
pillow
ultralytics
geopy
langgraph
langchain-google-genai
langchain-core
python-dotenv